# swagger/petstore.yml
swagger: "2.0"
info:
  title: "Swagger Petstore"
  version: "1.0.0"
basePath: /
paths:
...
  /store/order/:
    post:
      operationId: api.store.post_order
      parameters:
        - in: "body"
          name: "order"
          description: "order placed for purchasing the pet"
          required: true
          schema:
            $ref: '#/definitions/Order'
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/Order'
        400:
          description: "Invalid Order" 
        404:
          description: "Order not found"
...
definitions:
...
  Order:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      petId:
        type: "integer"
        format: "int64"
      quantity:
        type: "integer"
        format: "int32"
      shipDate:
        type: "string"
        format: "date-time"
      status:
        type: "string"
        description: "Order Status"
        enum:
          - "placed"
          - "approved"
          - "delivered"
      complete:
        type: "boolean"
        default: false
    xml:
      name: "Order"